Flytt modal ut av komponenten
